requirements.txt
Pillow==9.2.0
PyPDF2==2.5.0
pytube==12.1.0
